Refactorización de VentaService y VentaController

+++ src/main/java/com/proyectofinal/controller/controllers/VentaController.java
@@ -1,8 +1,10 @@
 package com.proyectofinal.controller.controllers;
 
-import com.proyectofinal.controller.dto.VentaAux;
 import com.proyectofinal.controller.dto.VentaDTO;
+import com.proyectofinal.entities.DetalleVenta;
+import com.proyectofinal.entities.Producto;
 import com.proyectofinal.entities.Venta;
+import com.proyectofinal.service.interfaces.IDetalleVentaService;
 import com.proyectofinal.service.interfaces.IVentaService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -21,6 +24,8 @@ public class VentaController {
 
     @Autowired
     private IVentaService ventaService;
+    @Autowired
+    private IDetalleVentaService detalleVentaService;
     private final ModelMapper modelMapper = new ModelMapper();
 
     @GetMapping("/find")
@@ -51,25 +56,79 @@ public class VentaController {
     }
 
     @PostMapping("/save")
-    public ResponseEntity<VentaDTO> save(@RequestBody VentaAux ventaAux) throws URISyntaxException {
+    public ResponseEntity<VentaDTO> save(@RequestBody VentaDTO ventaDTO) throws URISyntaxException {
 
-        this.ventaService.procesarVenta(ventaAux.getCliente(), ventaAux.getProductos());
+        Venta venta = Venta.builder()
+                .fecha_venta(LocalDate.now())
+                .cliente(ventaDTO.getCliente())
+                .listaProductos(ventaDTO.getListaProductos())
+                .build();
+
+        this.ventaService.validarStock(venta.getListaProductos());
+        double total = this.ventaService.calcularTotal(venta.getListaProductos());
+        this.ventaService.restarStock(venta.getListaProductos());
+        venta.setTotal(total);
+
+        this.ventaService.save(venta);
+
+        for(Producto producto : venta.getListaProductos()){
+
+            this.detalleVentaService.save(DetalleVenta.builder()
+                    .idVenta(venta.getCodigo_venta())
+                    .idProducto(producto.getCodigoProducto())
+                    .cantidad(producto.getCantidadSolicitada())
+                    .build()
+            );
+        }
 
         return ResponseEntity.created(new URI("/venta/save")).build();
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<VentaAux> update(@RequestBody VentaAux ventaAux, @PathVariable Long id){
+    public ResponseEntity<VentaDTO> update(@RequestBody VentaDTO ventaDTO, @PathVariable Long id){
 
-        this.ventaService.actulizarVenta(ventaAux.getCliente(), ventaAux.getProductos(), id);
+        Optional<Venta> ventaOptional = this.ventaService.findById(id);
 
-        return ResponseEntity.ok(ventaAux);
+        if(ventaOptional.isPresent()){
+            Venta venta = ventaOptional.get();
+            venta.setFecha_venta(LocalDate.now());
+            venta.setCliente(ventaDTO.getCliente());
+            venta.setListaProductos(ventaDTO.getListaProductos());
+            double total = this.ventaService.calcularTotal(venta.getListaProductos());
+            venta.setTotal(total);
+
+            this.ventaService.devolverStock(id);
+            this.ventaService.validarStock(venta.getListaProductos());
+            this.ventaService.restarStock(venta.getListaProductos());
+
+
+            for(Producto producto : venta.getListaProductos()){
+
+                this.detalleVentaService.save(DetalleVenta.builder()
+                        .idVenta(venta.getCodigo_venta())
+                        .idProducto(producto.getCodigoProducto())
+                        .cantidad(producto.getCantidadSolicitada())
+                        .build()
+                );
+            }
+
+            this.ventaService.save(venta);
+
+            return ResponseEntity.ok(ventaDTO);
+        }
+
+        return ResponseEntity.badRequest().build();
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<String> deleteById(@PathVariable Long id){
 
-        this.ventaService.deleteById(id);
+        Optional<Venta> ventaOptional = this.ventaService.findById(id);
+
+        if(ventaOptional.isPresent()){
+
+            this.ventaService.deleteById(id);
+        }
 
         return ResponseEntity.ok("Venta eliminada");
     }
